// src/demo/types.ts
export interface ChatMessage {
  /**
   * Message id
   * @type {number}
   * @memberof ChatMessage
   */
  id: number;
  /**
   * User id
   * @type {number}
   * @memberof ChatMessage
   */
  userId: number;
  /**
   * Chat id
   * @type {number}
   * @memberof ChatMessage
   */
  chatId: number;
  /**
   * Message sent time
   * @type {Date}
   * @memberof ChatMessage
   */
  time: Date;
  /**
   * Message type
   * @type {string}
   * @memberof ChatMessage
   */
  type: ChatMessage.TypeEnum;
  /**
   * Message content (message text for messages and resourceId for files)
   * @type {string}
   * @memberof ChatMessage
   */
  content: string;
  /**
   * File
   * @type {Resource}
   * @memberof ChatMessage
   */
  file?: Resource;
}

/**
 * @export
 * @namespace ChatMessage
 */
// eslint-disable-next-line @typescript-eslint/no-namespace
export namespace ChatMessage {
  /**
   * @export
   * @enum {string}
   */
  export enum TypeEnum {
    Message = <any>'message',
    File = <any>'file',
  }
}
// ...
export interface ChatsResponse {
  /**
   * Chat id
   * @type {number}
   * @memberof ChatsResponse
   */
  id: number;
  /**
   * Chat title
   * @type {string}
   * @memberof ChatsResponse
   */
  title: string;
  /**
   * Chat avatar
   * @type {string}
   * @memberof ChatsResponse
   */
  avatar: string;
  /**
   * Number of unread messages in chat for current user
   * @type {number}
   * @memberof ChatsResponse
   */
  unreadCount: number;
  /**
   *
   * @type {ChatsResponseLastMessage}
   * @memberof ChatsResponse
   */
  lastMessage: ChatsResponseLastMessage;
}

/**
 *
 * @export
 * @interface ChatsResponseLastMessage
 */
export interface ChatsResponseLastMessage {
  /**
   * Message user (sender)
   * @type {UserResponse}
   * @memberof ChatsResponseLastMessage
   */
  user?: UserResponse;
  /**
   * Message timestamp
   * @type {string}
   * @memberof ChatsResponseLastMessage
   */
  time?: string;
  /**
   * Message content
   * @type {string}
   * @memberof ChatsResponseLastMessage
   */
  content?: string;
}
// ...
export interface Resource {
  /**
   * Message id
   * @type {number}
   * @memberof Resource
   */
  id: number;
  /**
   * User id
   * @type {number}
   * @memberof Resource
   */
  userId: number;
  /**
   * Server relative file path
   * @type {string}
   * @memberof Resource
   */
  path: string;
  /**
   * Initial file name
   * @type {string}
   * @memberof Resource
   */
  filename: string;
  /**
   * File content type (e.g \"image/jpeg\" for .jpg images)
   * @type {string}
   * @memberof Resource
   */
  contentType: string;
  /**
   * File size in bytes
   * @type {number}
   * @memberof Resource
   */
  contentSize: number;
  /**
   * Resource uploading time
   * @type {Date}
   * @memberof Resource
   */
  uploadDate: Date;
}
// ...
export interface UserResponse {
  /**
   * User id
   * @type {number}
   * @memberof UserResponse
   */
  id: number;
  /**
   * First name
   * @type {string}
   * @memberof UserResponse
   */
  firstName: string;
  /**
   * Second name
   * @type {string}
   * @memberof UserResponse
   */
  secondName: string;
  /**
   * Display name
   * @type {string}
   * @memberof UserResponse
   */
  displayName: string;
  /**
   * User login - unique
   * @type {string}
   * @memberof UserResponse
   */
  login: string;
  /**
   * Email
   * @type {string}
   * @memberof UserResponse
   */
  email: string;
  /**
   * Phone
   * @type {string}
   * @memberof UserResponse
   */
  phone: string;
  /**
   * Avatar
   * @type {string}
   * @memberof UserResponse
   */
  avatar: string;
}